+++ src/functions/greetUser.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2021 The Welcome-Bot Team and Contributors
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
-const getGuild = require("../db/functions/getGuild");
+const getGuild = require("../db/functions/guild/getGuild");
 //const genImage = require("./genImage");
 //const fs = require("fs");
 module.exports = async (member) => {